fix: x64 prebuilt binaries get overridden by ia32 binaries during dependency rebuild
Close #3038
