+++ packages/electron-builder-lib/src/asar/asarUtil.ts
@@ -1,13 +1,12 @@
 import BluebirdPromise from "bluebird-lst"
 import { AsyncTaskManager, log } from "builder-util"
 import { FileCopier, Filter, MAX_FILE_REQUESTS } from "builder-util/out/fs"
-import { createReadStream, createWriteStream, ensureDir, readFile } from "fs-extra-p"
+import { createReadStream, createWriteStream, ensureDir, readFile, Stats, writeFile } from "fs-extra-p"
 import * as path from "path"
 import { AsarOptions } from ".."
 import { Packager } from "../packager"
 import { PlatformPackager } from "../platformPackager"
-import { copyFileOrData, getDestinationPath } from "../util/appFileCopier"
-import { ResolvedFileSet } from "../util/AppFileCopierHelper"
+import { ResolvedFileSet, getDestinationPath } from "../util/appFileCopier"
 import { AsarFilesystem, Node } from "./asar"
 import { detectUnpackedDirs } from "./unpackDetector"
 
@@ -246,4 +245,13 @@ async function order(filenames: Array<string>, orderingFile: string, src: string
   }
   log.info({coverage: ((total - missing) / total * 100)}, "ordering files in ASAR archive")
   return sortedFiles
+}
+
+function copyFileOrData(fileCopier: FileCopier, data: string | Buffer | undefined | null, source: string, destination: string, stats: Stats) {
+  if (data == null) {
+    return fileCopier.copy(source, destination, stats)
+  }
+  else {
+    return writeFile(destination, data)
+  }
 }
